requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6

# Serial Communication
pyserial==3.5

# Scientific Computing
numpy==1.24.3
scipy==1.11.4

# Point Cloud Processing
open3d==0.18.0

# MAVLink Integration
pymavlink==2.4.41

# Async Support
aiohttp==3.9.1
asyncio-mqtt==0.16.2

# WebSocket
python-socketio==5.10.0
websocket-client==1.6.4

# Data Storage
h5py==3.10.0

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
